[FIX] web: fix freezed screen when fetching unexisting record
Current behavior:
In the barcode app, if you create a new receipt and add a line with a
kit product. Then after you edit the line you just created, and put the
quantity to 0 then confirm, the screen will freeze and you will need to
refresh the page.

Steps to reproduce:
- Create a kit product with 2 component
- Open barcode app, create a new receipt
- Add the product, and validate the receipt (the kit will be split)
- Edit one line and set quantity to zero
- Confirm
- you are now stuck on this screen, confirm doesn't redirect

opw-3524228

closes odoo/odoo#150700

X-original-commit: 1672dd75c655ef97d3cb2427a9ded520c963c95a
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>